+++ houssg/src/helper/index.ts
@@ -15,10 +15,10 @@ import {
 import { getMyHouseListData, onRegiFunnelData } from './ownerFunction';
 
 // 객실 관련
-import { getTargetRoomData, addTargetRoom } from './roomFunction';
+import { getTargetRoomData, addTargetRoom, returnAddRoomFormData } from './roomFunction';
 
 export { authLoginFunc, authSignUpFunc, kakaoLoginFunc, kakaoSignUp, nickCheckFunc, idCheckFunc, onPhoneUsableCheck, phoneAuthCheck, onFindId };
 
 export { getMyHouseListData, onRegiFunnelData };
 
-export { getTargetRoomData, addTargetRoom };
+export { getTargetRoomData, addTargetRoom, returnAddRoomFormData };
