+++ houssg/src/types/index.ts
@@ -8,7 +8,7 @@ import { RegiStepProps, StepMoverType, FunnelPropsType } from './regi';
 import { AuthLoginFunc, AuthSignUpFunc } from './rtk';
 import { UserMyPageType } from './mypage';
 import { HouseBaseInfo, ServiceList } from './houseDetailPage';
-import { RoomDataType, RoomComp, RoomData } from './room';
+import { RoomDataType, RoomComp, RoomData, AddRoomProps } from './room';
 // auth Regi manage
 export type { AuthProps, AuthInputType, AuthModeType, AuthPropsWithState, IdFindingType, SmsParameter };
 
@@ -28,4 +28,4 @@ export type { UserMyPageType };
 export type { HouseBaseInfo, ServiceList };
 
 // room
-export type { RoomDataType, RoomComp, RoomData };
+export type { RoomDataType, RoomComp, RoomData, AddRoomProps };
